Validate on deploy, refactor and reorder logic for watching an app
